Add dependency gatsby-transformer-sharp
If `gatsby-source-wagtail` is installed into a relatively plain project
that does not have many other dependencies, it can easily happen that
the `gatsby-transformer-sharp` plugin is not installed.

This leads to errors, because `preview.boilerplate.js` contains the
following line (440):
```js
const fragmentTypes = require('gatsby-transformer-sharp/src/fragments.js')
```

The dependency on `gatsby-transformer-sharp` is now listed. The defined
version (2.5.6) is based on an example project I was working on a while
back. It would probably make sense to test lower versions of this
dependency as well.